Rename findpeaks to peakfind for consistency
