+++ photutils/detection/_utils.py
@@ -10,7 +10,7 @@ import warnings
 from astropy.stats import gaussian_fwhm_to_sigma
 import numpy as np
 
-from .findpeaks import find_peaks
+from .peakfind import find_peaks
 from ..utils._convolution import _filter_data
 from ..utils.exceptions import NoDetectionsWarning
