+++ photutils/detection/peakfind.py
